percentLift: unify on float (whole-percent) across every transport
Fixes the M5 docs-site sim's busted index bar (pinned at floor while
chevrons cycle) and unifies the percent-lift contract on every surface
that carries it.

## Why

PR #386 (3e5bc5a, v4.23 wire bump) widened the wire's percentLift
field to tenths-of-a-percent (0..999) and split the M5 firmware's
in-memory state into two correlated globals: `PercentLift` (int 0..99,
chevron color) and `PercentLiftDeci` (float 0.0..99.9, index-bar
y-position).  Both were populated inside `InjectSerialByte`.  The
DUMMY_SERIAL_DATA path in `SerialRead.cpp` (the docs-site WASM build)
never touched `PercentLiftDeci`, so the bar pinned at y=192 forever.
The bug was only visible in the docs-site sim — wire-fed paths
(production firmware over UART, X-Plane plugin, m5-replay) all funnel
through `InjectSerialByte` and got both globals set.

The fix isn't to add a missing assign — it's to delete the
two-globals trap.

## Changes

In-memory: `DisplayBuildInputs::percentLift` (int, tenths) and
`DisplayFrame::percentLift` (int, tenths) become `percentLiftPct`
(float, whole percent, 0.0..99.9).  `ComputePercentLift` returns
`float` in whole-percent units; `ComputePercentLiftTenths` is deleted.
The wire encoder still emits `%03u` tenths via `int(pct * 10.0f)`
truncation — wire bytes byte-for-byte identical to v4.23 master, no
coordinated reflash needed.

M5 firmware: `PercentLift` (was int) and `PercentLiftDeci` (was float)
collapse to a single `extern float PercentLift`.  Chevron / slip
comparisons against integer `Array[i]` continue via implicit float-int
promotion (exact for our [0, 99] range).  The on-screen 2-digit
readout rounds + clamps to [0, 99] so a saturated 99.9 doesn't render
as "100".  `drawAOA`'s `aoaPct` parameter widens from int to float so
sub-percent fidelity reaches `mapPct2Display`.  The dummy block writes
a float demoPct directly — same rendering path the wire path drives,
so a future drift would surface the same way on either path.

WebSocket JSON: `percentLift` field changes from `%i` integer to
`%.1f` float (e.g. `47.3`).  `tools/web/lib/ws/wsClient.js` parses
with parseFloat.  LiveView's index bar now reads at the same fidelity
the wire carries, and `pct2y.js` returns a fractional y so SVG
renders the bar with subpixel anti-aliasing — visibly smoother
motion than the M5 hardware (which has to round to int pixels
because M5GFX::fillRect only accepts ints).  `mapPct2DisplayInt`
is exported for tests / M5-parity checks.

X-Plane plugin: `DataRefAdapter` writes `in.percentLiftPct` from
`ComputePercentLift` directly (one line replaces the
ComputePercentLiftTenths callsite + the int-typed band-edge casts).

Python tools: `compute_percent_lift` returns float in whole-percent
units; `compute_percent_lift_tenths` is deleted.  `Frame.percent_lift`
(int tenths) becomes `Frame.percent_lift_pct` (float whole-percent);
`to_bytes` does the ×10 truncation at the encoding boundary.

Tests: `test_percent_lift` rewritten to assert float endpoints,
linearity, sub-percent fidelity, and the 99.9 saturation clamp that
keeps the wire encoder's truncation strictly < 1000.
`test_display_serial` and `test_display_pct_anchors` updated for the
type rename.  Geometry invariants gain a regression block that pins
the bug fix: pct=0.5 must produce y < 192, fractional pct=70.5 lands
strictly between pct=70 and pct=71 (sub-percent end-to-end).
`parse_frame.cpp` prints `percentLiftPct=%.1f` (was `%d` — would
have been a silent UB / -Werror break with the field rename).

## Test plan

```
pio test -e native                 # 994/994 pass
pio run -e esp32s3-v4p             # SUCCESS
pio run -e m5stack-core-esp32 \
        -e m5stack-core2 \
        -e huvver-avi \
        -d software/OnSpeed-M5-Display    # 3/3 SUCCESS
cd software/OnSpeed-XPlane-Plugin && \
  cmake -B build -S . -DENABLE_M5_INDEXER=1 && \
  cmake --build build -j           # SUCCESS
software/OnSpeed-M5-Display/sim/build_wasm.sh   # SUCCESS
node tools/web/test/geometry-invariants.mjs     # 59 pass, 0 fail
uv run --with pytest --with pyserial python3 -m pytest \
       tools/onspeed_py/tests/ tools/m5-replay/test_replay.py
                                   # 50 passed
```

Visually verified in the rebuilt WASM sim: the white index bar moves
through the full sweep on every frame, tracking the AOA digit
correctly (top of column at PctLift=91, just above floor at PctLift=06).

+++ software/Libraries/onspeed_core/src/aoa/DisplayPctAnchors.cpp
@@ -59,10 +59,15 @@ void FillSnappedOperational(DisplayPctAnchors& out,
                             const SuFlaps& active,
                             bool iasValid)
 {
-    out.tonesOnPctLift     = ComputePercentLift(active.fLDMAXAOA,        active, iasValid);
-    out.onSpeedFastPctLift = ComputePercentLift(active.fONSPEEDFASTAOA,  active, iasValid);
-    out.onSpeedSlowPctLift = ComputePercentLift(active.fONSPEEDSLOWAOA,  active, iasValid);
-    out.stallWarnPctLift   = ComputePercentLift(active.fSTALLWARNAOA,    active, iasValid);
+    // ComputePercentLift returns float (whole percent); the snapped
+    // band-edge fields stay integer-percent on the wire (only move on
+    // detent / config-save events, sub-percent buys nothing). Truncate
+    // matches `static_cast<int>(ComputePercentLift(...))` everywhere
+    // these snapped anchors are surfaced.
+    out.tonesOnPctLift     = static_cast<int>(ComputePercentLift(active.fLDMAXAOA,        active, iasValid));
+    out.onSpeedFastPctLift = static_cast<int>(ComputePercentLift(active.fONSPEEDFASTAOA,  active, iasValid));
+    out.onSpeedSlowPctLift = static_cast<int>(ComputePercentLift(active.fONSPEEDSLOWAOA,  active, iasValid));
+    out.stallWarnPctLift   = static_cast<int>(ComputePercentLift(active.fSTALLWARNAOA,    active, iasValid));
 }
 
 // Compute the visual pip's "full-flap target" percent: the bottom-half
@@ -79,8 +84,8 @@ void FillSnappedOperational(DisplayPctAnchors& out,
 // the upper donut to meet a band-center pip.
 int FullFlapPipTarget(const SuFlaps& mostDeployed, bool iasValid)
 {
-    const int fastPct = ComputePercentLift(mostDeployed.fONSPEEDFASTAOA, mostDeployed, iasValid);
-    const int slowPct = ComputePercentLift(mostDeployed.fONSPEEDSLOWAOA, mostDeployed, iasValid);
+    const int fastPct = static_cast<int>(ComputePercentLift(mostDeployed.fONSPEEDFASTAOA, mostDeployed, iasValid));
+    const int slowPct = static_cast<int>(ComputePercentLift(mostDeployed.fONSPEEDSLOWAOA, mostDeployed, iasValid));
     int target = static_cast<int>(std::lround(
         (3.0f * static_cast<float>(fastPct) + static_cast<float>(slowPct)) / 4.0f));
     if (target < 0)  target = 0;
@@ -141,7 +146,7 @@ DisplayPctAnchors ComputeDisplayPctAnchors(
         const int potB      = mostDeployed.iPotPosition;
         const int span      = potB - potA;
 
-        const int pipClean    = ComputePercentLift(cleanest.fLDMAXAOA, cleanest, iasValid);
+        const int pipClean    = static_cast<int>(ComputePercentLift(cleanest.fLDMAXAOA, cleanest, iasValid));
         const int pipFullFlap = FullFlapPipTarget(mostDeployed, iasValid);
 
         float lambda;
